Moved adapter to main activity
Left fragment pantry empty since adapter only worked properly with listview in an activity, will have to adjust accordingly. This provides list options for various other tasks
