+++ app/src/main/java/com/mychef/mychef/PantryFragment.kt
@@ -7,30 +7,11 @@ import android.widget.ListView
 import androidx.fragment.app.Fragment
 import android.os.Bundle
 import android.R
+import android.content.Context
 
-class PantryFragment: Fragment(R.layout.fragment_pantry) {
-    private class CustomAdapter(context: Context): BaseAdapter {
+class PantryFragment: Fragment() {
 
-        private val mContext: Context
 
-        init {
-            this.mContext = context
-        }
-        fun getCount(): Int {
-
-        }
-        fun getItemId(position: Int): Long{
-            return position.toLong()
-        }
-        fun getItem(position: Int): Any{
-            return "TESTING"
-        }
-        fun getView(position: Int, convertView: View?, viewGroup: ViewGroup?): View {
-            val textView = TextView(mContext)
-            textView.text = "Test Row for ListView"
-            return textView
-        }
-    }
 }
